requirements.txt
bleach==6.1.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-paginate==2024.4.12
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.3
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
